Code re-orgainzation.
Added clgetValp.h which has the API-level templated functions which
use the BaseCode functions from clgetBaseCode.h

clgetBaseCode.h has only the various BaseCode functions.

tstcpp.cc include clgetValp.h (not clgetBaseCode.h which has the lower-level interface).
All application code should include clgetValp.h
