+++ code/tstcpp.cc
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include <vector>
 #include <exception>
-#include <clgetBaseCode.h>
+#include <clgetValp.h>
 /*
    Test program to test the embedded shell via the commandline library
 */
